% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADR_function.R
\name{ADR_function}
\alias{ADR_function}
\title{ADR_function}
\usage{
ADR_function(high = "high", low = "low", days = 20)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{days}{days}
}
\value{
ADR
}
\description{
ADR_function
}
