% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_possible_numeric.R
\name{check_possible_numeric}
\alias{check_possible_numeric}
\title{check_possible_numeric}
\usage{
check_possible_numeric(data)
}
\arguments{
\item{data}{data}
}
\value{
vector index
}
\description{
check_possible_numeric
}
