% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_yf.R
\name{convert_to_yf}
\alias{convert_to_yf}
\title{convert_to_yf}
\usage{
convert_to_yf(symbol, exchange = exchange)
}
\arguments{
\item{symbol}{symbol}

\item{exchange}{exchange}
}
\value{
ticker
}
\description{
convert_to_yf
}
