% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_1year_return_for_all.R
\name{fetch_1year_return_for_all}
\alias{fetch_1year_return_for_all}
\title{fetch_1year_return_for_all}
\usage{
fetch_1year_return_for_all(key = Sys.getenv("BORSDATA_KEY"))
}
\arguments{
\item{key}{API key}
}
\value{
kurs data frame
}
\description{
fetch_1year_return_for_all
}
