% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_date_price_for_all.R
\name{fetch_date_price_for_all}
\alias{fetch_date_price_for_all}
\title{fetch_date_price_for_all}
\usage{
fetch_date_price_for_all(DATE = "2022-04-05", key = Sys.getenv("BORSDATA_KEY"))
}
\arguments{
\item{DATE}{date}

\item{key}{API key}
}
\value{
kurs data frame
}
\description{
fetch_date_price_for_all
}
