% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_instruments.R
\name{fetch_instruments}
\alias{fetch_instruments}
\title{fetch_instruments}
\usage{
fetch_instruments(key = key)
}
\arguments{
\item{key}{API key}
}
\value{
df data frame
}
\description{
fetch_instruments
}
