% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_multiple_stock.R
\name{fetch_multiple_stock}
\alias{fetch_multiple_stock}
\title{fetch_multiple_stock}
\usage{
fetch_multiple_stock(df)
}
\arguments{
\item{df}{a subset of companies}
}
\value{
df data frame
}
\description{
fetch_multiple_stock
}
