% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_quarter.R
\name{fetch_quarter}
\alias{fetch_quarter}
\title{fetch_quarter}
\usage{
fetch_quarter(id, key = Sys.getenv("BORSDATA_KEY"))
}
\arguments{
\item{id}{ins id}

\item{key}{API key}
}
\value{
df data frame
}
\description{
fetch_quarter
}
