% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rsi_for_all.R
\name{fetch_rsi_for_all}
\alias{fetch_rsi_for_all}
\title{fetch_rsi_for_all}
\usage{
fetch_rsi_for_all(days = 3, key = Sys.getenv("BORSDATA_KEY"))
}
\arguments{
\item{days}{number of days}

\item{key}{API key}
}
\value{
kurs data frame
}
\description{
fetch_rsi_for_all
}
