% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sma_diff_day_for_all.R
\name{fetch_sma_diff_day_for_all}
\alias{fetch_sma_diff_day_for_all}
\title{fetch_sma_diff_day_for_all}
\usage{
fetch_sma_diff_day_for_all(days = 50, key = Sys.getenv("BORSDATA_KEY"))
}
\arguments{
\item{days}{days}

\item{key}{API key}
}
\value{
kurs data frame
}
\description{
fetch_sma_diff_day_for_all
}
