% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_stockprice.R
\name{fetch_stockprice}
\alias{fetch_stockprice}
\title{fetch_stockprice}
\usage{
fetch_stockprice(id, since = "2018-01-01", key = Sys.getenv("BORSDATA_KEY"))
}
\arguments{
\item{id}{Instrument id}

\item{since}{A date}

\item{key}{API key}
}
\value{
kurs data frame
}
\description{
fetch_stockprice
}
