% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnet_plot.R
\name{fishnet_plot}
\alias{fishnet_plot}
\title{fishnet_plot}
\usage{
fishnet_plot(
  symbol = "^GSPC",
  start_date = "2021-01-01",
  end_date = lubridate::today(),
  save_plot = TRUE,
  label_size = 180,
  show_panel = FALSE
)
}
\arguments{
\item{symbol}{ticker}

\item{start_date}{start_date}

\item{end_date}{end_date}

\item{save_plot}{save_plot T/F}

\item{label_size}{label_size}

\item{show_panel}{show_panel T/F}
}
\value{
p
}
\description{
fishnet_plot
}
