% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators}
\alias{indicators}
\title{indicators}
\usage{
indicators(data, high = high, low = low, close = close, open = open)
}
\arguments{
\item{data}{data}

\item{high}{high}

\item{low}{low}

\item{close}{close}

\item{open}{open}
}
\value{
df
}
\description{
indicators
}
