% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsalomon-package.R
\docType{package}
\name{jsalomon-package}
\alias{jsalomon}
\alias{jsalomon-package}
\title{jsalomon: Market analysis}
\description{
This is a personal package for trading and market analysis.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
