% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_and_vallies_plot.R
\name{peaks_and_vallies_plot}
\alias{peaks_and_vallies_plot}
\title{peaks_and_vallies_plot}
\usage{
peaks_and_vallies_plot(ticker = "^GSPC", start_date = "2021-01-01", M = 12)
}
\arguments{
\item{ticker}{ticker}

\item{start_date}{start_date}

\item{M}{size}
}
\value{
p
}
\description{
peaks_and_vallies_plot
}
