% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_table.R
\name{perf_table}
\alias{perf_table}
\title{perf_table}
\usage{
perf_table(country = "sweden", exchange = "OMXSTO")
}
\arguments{
\item{country}{country sweden america}

\item{exchange}{exchange like OMXSTO}
}
\value{
z tibble
}
\description{
perf_table
}
