% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_table1.R
\name{perf_table1}
\alias{perf_table1}
\title{perf_table1}
\usage{
perf_table1(country = "sweden", exchange = "OMXSTO")
}
\arguments{
\item{country}{country sweden america}

\item{exchange}{exchange like OMXSTO}
}
\value{
z tibble
}
\description{
perf_table1
}
