% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_table_common_ETF.R
\name{perf_table_common_ETF}
\alias{perf_table_common_ETF}
\title{perf_table_common_ETF}
\usage{
perf_table_common_ETF(
  country = "america",
  exchange = "AMEX",
  valueTraded = 1e+07
)
}
\arguments{
\item{country}{country sweden america}

\item{exchange}{exchange like OMXSTO}

\item{valueTraded}{valueTraded}
}
\value{
z tibble
}
\description{
perf_table_common_ETF
}
