% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_table_multi.R
\name{perf_table_multi}
\alias{perf_table_multi}
\title{perf_table_multi}
\usage{
perf_table_multi(country = "america")
}
\arguments{
\item{country}{nordic america}
}
\value{
screen_data screen_data
}
\description{
perf_table_multi
}
