% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_table_multi_robust.R
\name{perf_table_multi_robust}
\alias{perf_table_multi_robust}
\title{perf_table_multi_robust}
\usage{
perf_table_multi_robust(country = "america")
}
\arguments{
\item{country}{nordic america}
}
\value{
screen_data screen_data
}
\description{
perf_table_multi_robust
}
