% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock.R
\name{plot_stock}
\alias{plot_stock}
\title{ADR_function}
\usage{
plot_stock(ticker, plot_h = 300)
}
\arguments{
\item{ticker}{ticker}

\item{plot_h}{lookback period in days}
}
\value{
p
}
\description{
ADR_function
}
