% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock3.R
\name{plot_stock3}
\alias{plot_stock3}
\title{plot_stock3 test}
\usage{
plot_stock3(ticker, plot_h = 350, zoom_days = 55)
}
\arguments{
\item{ticker}{ticker}

\item{plot_h}{lookback period in days}

\item{zoom_days}{zoom days}
}
\value{
p
}
\description{
plot_stock3 test
}
