% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock4.R
\name{plot_stock4}
\alias{plot_stock4}
\title{plot_stock4 test}
\usage{
plot_stock4(ticker, plot_h = 350, zoom_days = 55, title_p = TRUE)
}
\arguments{
\item{ticker}{ticker}

\item{plot_h}{lookback period in days}

\item{zoom_days}{zoom days}

\item{title_p}{plot titles}
}
\value{
p
}
\description{
plot_stock4 test
}
