% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_size.R
\name{position_size}
\alias{position_size}
\title{position_size}
\usage{
position_size(risk_in_money, price_entry, stoploss)
}
\arguments{
\item{risk_in_money}{risk_in_money}

\item{price_entry}{price_entry}

\item{stoploss}{stoploss}
}
\value{
df
}
\description{
position_size
}
