% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS_function.R
\name{RS_function}
\alias{RS_function}
\title{RS_function}
\usage{
RS_function(tv_data, col = perf_ytd, industry_rank = FALSE)
}
\arguments{
\item{tv_data}{tv_data}

\item{col}{performance column}

\item{industry_rank}{t/F}
}
\value{
data
}
\description{
RS_function
}
