% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_salo.R
\name{scale_color_salo}
\alias{scale_color_salo}
\title{scale_color_salo}
\usage{
scale_color_salo(
  theme = "JS",
  key = list(JS = c("#4D6EDD", "#5FFB17", "#FF4D6B", "#FFD800", "#00E5E5", "#FF6C02",
    "#B47CFF", "#FF2020"))
)
}
\arguments{
\item{theme}{theme JS}

\item{key}{list of colors}
}
\description{
scale_color_salo
}
