% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bors.R
\name{theme_bors}
\alias{theme_bors}
\title{theme_bors}
\usage{
theme_bors(base_size = 2.5, plotbq = "#040926", legendposition = "none")
}
\arguments{
\item{base_size}{base_size}

\item{plotbq}{plot background color}

\item{legendposition}{legend position}
}
\value{
A theme element
}
\description{
theme_bors
}
