% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock2.R
\name{plot_stock2}
\alias{plot_stock2}
\title{plot_stock2}
\usage{
plot_stock2(ticker, plot_h = 350, zoom_days = 55)
}
\arguments{
\item{ticker}{ticker}

\item{plot_h}{lookback period in days}

\item{zoom_days}{zoom days}
}
\value{
p
}
\description{
plot_stock2
}
