% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock5.R
\name{plot_stock5}
\alias{plot_stock5}
\title{plot_stock5 test}
\usage{
plot_stock5(
  ticker,
  plot_h = 400,
  zoom_days = 55,
  title_p = FALSE,
  pos_col = "#1ED01E",
  neg_col = "#DC0101",
  tight_col = "white"
)
}
\arguments{
\item{ticker}{ticker}

\item{plot_h}{lookback period in days}

\item{zoom_days}{zoom days}

\item{title_p}{plot titles}

\item{pos_col}{col}

\item{neg_col}{col}

\item{tight_col}{col}
}
\value{
p
}
\description{
plot_stock5 test
}
